// src/modules/balancer-sdk/sor/impl/types.ts
export interface SubgraphPoolBase {
  id: string;
  address: string;
  poolType: string;
  swapFee: string;
  swapEnabled: boolean;
  totalShares: string;
  tokens: SubgraphToken[];
  tokensList: string[];

  // Weighted & Element field
  totalWeight?: string;

  // Stable specific fields
  amp?: string;

  // Element specific fields
  expiryTime?: number;
  unitSeconds?: number;
  principalToken?: string;
  baseToken?: string;

  // Linear specific fields
  mainIndex?: number;
  wrappedIndex?: number;
  lowerTarget?: string;
  upperTarget?: string;

  // Gyro2 specific fields
  sqrtAlpha?: string;
  sqrtBeta?: string;

  // Gyro3 specific field
  root3Alpha?: string;

  // GyroE specific fields
  alpha?: string;
  beta?: string;
  c?: string;
  s?: string;
  lambda?: string;
  tauAlphaX?: string;
  tauAlphaY?: string;
  tauBetaX?: string;
  tauBetaY?: string;
  u?: string;
  v?: string;
  w?: string;
  z?: string;
  dSq?: string;
}

export type SubgraphToken = {
  address: string;
  balance: string;
  decimals: number;
  priceRate: string;
  // WeightedPool field
  weight: string | null;
};
// ...
export interface PoolDataService {
  getPools(): Promise<SubgraphPoolBase[]>;
}